% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_subnetwork.R
\name{create_subnetwork}
\alias{create_subnetwork}
\title{Creates a subnetwork from a COSIFER network}
\usage{
create_subnetwork(
  directory,
  outDirectory,
  idx = NULL,
  threshold = NULL,
  n.edges = 10^5
)
}
\arguments{
\item{outDirectory}{Path to the output directory}

\item{idx}{Optional: index of the network file when multiple are present}

\item{file}{Path to the input file}
}
\value{
Stores the subnetwork and its largest connected component into different formats for further use
}
\description{
This function loads a network file produced using COSIFER ( with extension .csv.gz) as a dataframe.
It takes the network file path directory as an input, and stores the subnetwork in the output directory.
If multiple .csv.gz files are present, the user can select a specific file using the idx parameter.
Additional parameters are passed to the function "filter_edges" for the creation of the subnetwork.
The function stores the subnetwork, the largest connected component of the network as well as
its coexpression matrix as an R session (.Rdata extension); the largest connected component edge weighet dataframe
in a .csv.gz file and as an igraph object in a .sif format (also to easy visualization into cytoscape)
}
