% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_functions.R
\name{filter_edges}
\alias{filter_edges}
\title{Filter/remove edges from a network}
\usage{
filter_edges(df.net, w.threshold = NULL, n.edges = 10^5)
}
\arguments{
\item{df.net}{An edge weighted dataframe}

\item{w.threshold}{A weight threshold}

\item{n.edges}{A number of edges to retain}
}
\value{
A reduced edge weighted dataframe
}
\description{
The function takes as input an edge weighted dataframe encoded in its most typical form:
with column names "from", "to", "weight".
It filters and removes edges either based on a threshold (w.threshold)
either based on the number of edges to keep (n.edges),
selecting for the ones that have the highest weights; and
it returns the reduced edge weighted dataframe.
}
