% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_functions.R
\name{plot_density}
\alias{plot_density}
\title{Plot edge weight density for two edge weighted dataframes}
\usage{
plot_density(
  df.net,
  df.sub.net,
  filename = "density",
  dir = "",
  main.1 = "density",
  main.2 = "reduced network"
)
}
\arguments{
\item{df.net}{An edge weighted dataframe}

\item{df.sub.net}{An edge weighteddataframe}

\item{filename}{A pdf filename}

\item{dir}{An output directory}

\item{main.1}{A title for the first density plot}

\item{main.2}{A title for the second density plot}
}
\value{
A PDF plot
}
\description{
The function takes as input two edge weighted dataframes encoded in its most typical form:
with column names "from", "to", "weight". It produces the edge weight density plots
one beside the other and store it as a pdf file to the given directory.
}
