% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_functions.R
\name{plot_network}
\alias{plot_network}
\title{Plot a network}
\usage{
plot_network(net, title = NULL)
}
\arguments{
\item{net}{An igraph network}

\item{title}{A title for the plot}
}
\value{
A network plot
}
\description{
The function takes as input an igraph network.
If communities are present as vertex attributes of the network
nodes are colored based on the community they belong to.
If communities are not present, nodes are colored using the default igraph color.
}
