% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_functions.R
\name{read_big_files}
\alias{read_big_files}
\title{Load a big network}
\usage{
read_big_files(file)
}
\arguments{
\item{file}{Path to the input file}
}
\value{
A dataframe of the file
}
\description{
This function loads network files produced using COSIFER as a matrix.
The column names of such networks are "e1", "e2", "intensity".
The function uses this characteristic to import column by column the network.
Finally, columns are concatenated and a dataframe is returned.
}
