% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gprofiler.R
\name{run_gprofiler}
\alias{run_gprofiler}
\title{Performs enrichment analysis}
\usage{
run_gprofiler(directory)
}
\arguments{
\item{directory}{Path to the .Rdata object}
}
\value{
Stores the enrichment results in .csv and .Rdata objets
}
\description{
This function loads the Rdata session produced using the mcl_clustering function:
There mcl clustering was performed and stored as vertex attribute 'community' of an igraph object.
For each community (or set of nodes), the function performs gene set enrichment analysis using gprofiler2 package.
Results are stored in a .csv file with columns cluster number, function, and gene set.
The detailed gprofiler results is stored as a list in a Rdata working session.
}
