% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runRoma.R
\name{runRoma}
\alias{runRoma}
\title{Computes the eigenGene matrix using rROMA}
\usage{
runRoma(directory, outDirectory)
}
\arguments{
\item{directory}{Path to the expression matrix}

\item{outDirectory}{Path to the .RData gene set list}
}
\value{
Stores the eigenGene table as a .csv objet
}
\description{
This function loads the expression matrix in the input directory, and the .Rdata object
produced by the run_gprofiler function, storing a list of gene sets used by rRoma.R function of
the rRoma package.
It stores the eigenGene table as a .csv file at the input directory.
}
