% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_functions.R
\name{tunning.inflation}
\alias{tunning.inflation}
\title{Tunes the inflation parameter of MCL clustering}
\usage{
tunning.inflation(file, outDirectory)
}
\arguments{
\item{file}{An input directory}

\item{outDirectory}{An output directory}
}
\value{
A list with the mcl.clustering results and the number of communities with more than 10 genes
}
\description{
The function takes as input the directory of a R working space
where a coexpression matrix of a network has been stored;
and an output directory where to store the clustering results
computed with different values of the inflation parameter.
Selection of the inflation parameter is performed selecting for the value
leading to the maximum number of clusters with more than 10 nodes.
The function returns a list with the MCL clustering for the selected
inflation parameter and the number of clusters with more than 10 nodes
}
